// src/app/api/models/send-email-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { EmailAttachmentDto } from '../models/email-attachment-dto';
export interface SendEmailDto {
  attachments?: Array<EmailAttachmentDto>;
  bcc?: Array<string>;
  cc?: Array<string>;
  htmlContent?: string;
  priority?: 'low' | 'normal' | 'high';
  subject: string;
  textContent?: string;
  to: Array<string>;
  trackClicks?: boolean;
  trackOpens?: boolean;
}